+++ src/main/java/com/example/Scrabble/Model/Player.java
@@ -1,0 +1,12 @@
+package com.example.Scrabble.Model;
+
+public interface Player {
+    public String getName();
+    public void setName(String name);
+    public int getPlayerID();
+    public void setPlayerID(int playerID);
+    public int getScore();
+    public void setScore(int score);
+
+
+}
