+++ src/main/java/com/example/Scrabble/Model/PlayerHandler.java
@@ -4,18 +4,23 @@ import com.example.Scrabble.ScrabbleServer.ClientHandler;
 
 import java.io.*;
 
-public class ModelHandler implements ClientHandler {
+public class PlayerHandler implements ClientHandler {
 
+    GameManager pm;
     BufferedReader in;
     PrintWriter out;
     @Override
     public void handleClient(InputStream inFromclient, OutputStream outToClient) {
+        pm = GameManager.get();
         try {
             in = new BufferedReader(new InputStreamReader(inFromclient));
             out = new PrintWriter(outToClient, true);
             String line = in.readLine();
-            String newline = new String(line + " from server");
-            out.println(newline);
+
+
+            //logic for handling the model requests goes here (i.e. the logic for the model)
+            //System.out.println("this came from the handleClient: "+ line);
+            out.println("from: ModelHandler");
         }
         catch (IOException e){
             e.printStackTrace();
