+++ src/main/java/com/example/Scrabble/Test/TestMain.java
@@ -1,7 +1,9 @@
 package com.example.Scrabble.Test;
 
 
-import com.example.Scrabble.Model.HostModel;
+import com.example.Scrabble.Model.GuestPlayer;
+import com.example.Scrabble.tmp.HostModel;
+import com.example.Scrabble.Model.HostPlayer;
 
 // import java.util.Random;
 
@@ -13,9 +15,14 @@ public class TestMain {
         // Random r=new Random();
         // int port=6000+r.nextInt(1000);
         h.connectServer();
-
+        System.out.println("Server connected");
         h.test("waka");
-        h.BookActions("Q,s1.txt,s2.txt,19393");
+        h.dictionaryRequest("Q,s1.txt,s2.txt,19393");
+        GuestPlayer gp = new GuestPlayer();
+        System.out.println("Guest created " + gp);
+        HostPlayer hp = new HostPlayer();
+        System.out.println("Host created " + hp.getName());
+
